+++ opentelemetry-sdk/tests/resources/test_resources.py
@@ -238,9 +238,6 @@ class TestResources(unittest.TestCase):
                 {
                     SERVICE_NAME: "test",
                     "non-primitive-data-type": {},
-                    "invalid-byte-type-attribute": (
-                        b"\xd8\xe1\xb7\xeb\xa8\xe5 \xd2\xb7\xe1"
-                    ),
                     "": "empty-key-value",
                     None: "null-key-value",
                     "another-non-primitive": uuid.uuid4(),